// src/typings/itemsadder/schemas.d.ts
export interface IaSchema {
  info: {
    namespace: string;
    "dictionary-lang"?: string;
    [k: string]: unknown;
  };
  items?: {
    [k: string]: Item;
  };
  recipes?: {
    /**
     * Crafting table recipes
     */
    crafting_table?: {
      [k: string]: RecipeCraft;
    };
    cooking?: {
      [k: string]: RecipeCooking;
    };
    anvil_repair?: {
      [k: string]: RecipeAnvilRepair;
    };
    smithing?: {
      [k: string]: RecipeSmithing;
    };
    [k: string]: unknown;
  };
  /**
   * Create loots (item drop) based on broken block, killed entity and on fishing
   */
  loots?: {
    blocks?: {
      [k: string]: LootBlock;
    };
    mobs?: {
      [k: string]: LootMobs;
    };
    fishing?: {
      [k: string]: LootFishing;
    };
    [k: string]: unknown;
  };
  /**
   * Create rules to spawn custom ores around your worlds.
   * **They will be spawned only on new chunks.**
   * **Pregenerated worlds or already generated chunks won't be affected to avoid risky overwrite**
   */
  worlds_populators?: {
    [k: string]: WorldsPopulator;
  };
  /**
   * Create rules to spawn custom trees around your worlds.
   * **They will be spawned only on new chunks.**
   * **Pregenerated worlds or already generated chunks won't be affected to avoid risky overwrite**
   */
  trees_populators?: {
    [k: string]: TreesPopulators;
  };
  /**
   * Create rules to spawn custom blocks on the surface of your worlds.
   * **They will be spawned only on new chunks.**
   * **Pregenerated worlds or already generated chunks won't be affected to avoid risky overwrite**
   */
  surface_decorators?: {
    [k: string]: SurfaceDecorators;
  };
  /**
   * Create rules to spawn custom blocks in the caves of your worlds.
   * **They will be spawned only on new chunks.**
   * **Pregenerated worlds or already generated chunks won't be affected to avoid risky overwrite**
   */
  cave_decorators?: {
    [k: string]: CaveDecorators;
  };
  /**
   * Create your own font characters. You can use them as emoji, GUI texture, HUD...
   * It's basically an image shown in texts.
   */
  font_images?: {
    [k: string]: FontImage;
  };
  /**
   * Create your own HUDs. You will be able to show mana on screen, fuel, current minigame level, heat and anything you want to add!
   */
  huds?: {
    [k: string]: Hud;
  };
  /**
   * Create your own categories and show them in /ia command
   */
  categories?: {
    [k: string]: IaCategory;
  };
  /**
   * Create your own colored liquids
   */
  liquids?: {
    [k: string]: CustomLiquid;
  };
  /**
   * Replace language attributes to your custom value.
   */
  minecraft_lang_overwrite?: {
    [k: string]: MinecraftLangOverwrite;
  };
  /**
   * Create a special list of texts to be used (almost) anywhere and get translated automatically by the plugin.
   * For example: in items names, lores, /ia menu categories names
   */
  dictionary?: {
    "display-category-EXAMPLE"?: string;
    "display-category-EXAMPLE_2"?: string;
    "display-name-EXAMPLE_ITEM"?: string;
    "display-lore-EXAMPLE_ITEM"?: string;
    EXAMPLE?: string;
    [k: string]: string;
  };
  /**
   * Create a custom entity with 3D model
   */
  entities?: {
    [k: string]: Entity;
  };
  emotes?: {
    emote?: Emote;
    emote_2?: Emote;
    emote_3?: Emote;
    [k: string]: Emote1;
  };
  armors_rendering?: ArmorsRendering;
  [k: string]: unknown;
}
/**
 * Custom item
 */
export interface Item {
  enabled?: boolean;
  template?: boolean;
  variant_of?: string;
  display_name: string;
  permission_suffix?: string;
  /**
   * Special property to mark item as MMOITEM https://www.spigotmc.org/resources/39267/
   */
  mmoitem?: {
    type?: string;
    id?: string;
    use_mmoitem_displayname?: string;
  };
  /**
   * Resource pack model generator settings.
   * `material` property is required for every item.
   * However custom armors do not require this setting.
   * Custom armors default item material will be leather if not specified.
   */
  resource: {
    material?: VanillaMaterials;
    generate?: boolean;
    model_id?: number;
    model_path?: string;
    [k: string]: unknown;
  };
  lore?: string[];
  /**
   * Slot in which the item has effect
   */
  attribute_modifiers?: {
    mainhand?: AttributeModifiers;
    offhand?: AttributeModifiers;
    head?: AttributeModifiers;
    chest?: AttributeModifiers;
    legs?: AttributeModifiers;
    feet?: AttributeModifiers;
    [k: string]: unknown;
  };
  /**
   * Durability of the item
   */
  durability?: {
    max_custom_durability?: number;
    custom_durability?: number;
    disappear_when_broken?: boolean;
    unbreakable?: boolean;
    usages?: number;
    show_custom_durability_lore?: boolean;
    [k: string]: unknown;
  };
  item_flags?: (
    | "HIDE_ATTRIBUTES"
    | "HIDE_DESTROYS"
    | "HIDE_ENCHANTS"
    | "HIDE_PLACED_ON"
    | "HIDE_POTION_EFFECTS"
    | "HIDE_UNBREAKABLE"
  )[];
  enchants?: (VanillaEnchants | string)[];
  blocked_enchants?: (VanillaEnchants | "ALL")[];
  specific_properties?: SpecificProperties;
  nbt?: string;
  /**
   * Customize some aspects of how the dropped item appears.
   */
  drop?: {
    show_name?: boolean;
    /**
     * You can make an item glowing when dropped on the ground.
     * Very useful for rare items.
     */
    glow?: {
      enabled?: boolean;
      color?:
        | "BLACK"
        | "DARK_BLUE"
        | "DARK_GREEN"
        | "DARK_AQUA"
        | "DARK_RED"
        | "DARK_PURPLE"
        | "GOLD"
        | "GRAY"
        | "DARK_GRAY"
        | "BLUE"
        | "GREEN"
        | "AQUA"
        | "RED"
        | "PURPLE"
        | "YELLOW"
        | "WHITE"
        | "NONE";
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  behaviours?: Behaviours;
  events_cooldown?: number;
  /**
   * Special settings for events.
   */
  events_settings?: {
    /**
     * Cooldown settings for this item.
     */
    cooldown?: {
      ticks?: number;
      indicator?: "HIDDEN" | "TITLE" | "BOSSBAR" | "HUD";
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  all_actions_permission?: string;
  events?: Events;
  permission?: string;
  [k: string]: unknown;
}
// ...
export interface RecipeCraft {
  enabled?: boolean;
  permission_suffix?: string;
  shapeless?: boolean;
  ingredients: Ingredients;
  result: {
    item: VanillaItemsAdderCustomItem71;
    amount?: number;
    [k: string]: unknown;
  };
  return_items?: {
    decrement_durability?: {
      [k: string]: {
        item?: VanillaItemsAdderCustomItem72;
        amount?: number;
        [k: string]: unknown;
      };
    };
    increment_durability?: {
      [k: string]: {
        item?: VanillaItemsAdderCustomItem73;
        amount?: number;
        [k: string]: unknown;
      };
    };
    replace?: {
      [k: string]: VanillaItemsAdderCustomItem74;
    };
    play_sound?: {
      name?: VanillaSoundsAndCustom1 & VanillaSoundsAndCustom2;
      volume?: number;
      pitch?: number;
      category?: VanillaSoundCategory;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
// ...
export interface RecipeCooking {
  enabled?: boolean;
  permission_suffix?: string;
  machines: ("FURNACE" | "BLAST_FURNACE" | "SMOKER")[];
  ingredient: {
    item: VanillaItemsAdderCustomItem77;
    [k: string]: unknown;
  };
  exp?: number;
  cook_time?: number;
  result: {
    item: VanillaItemsAdderCustomItem80;
    amount?: number;
    return_items?: {
      decrement_durability?: {
        [k: string]: {
          item?: VanillaItemsAdderCustomItem81;
          amount?: number;
          [k: string]: unknown;
        };
      };
      increment_durability?: {
        [k: string]: {
          item?: VanillaItemsAdderCustomItem82;
          amount?: number;
          [k: string]: unknown;
        };
      };
      replace?: {
        [k: string]: VanillaItemsAdderCustomItem83;
      };
      play_sound?: {
        name?: VanillaSoundsAndCustom1 & VanillaSoundsAndCustom2;
        volume?: number;
        pitch?: number;
        category?: VanillaSoundCategory;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface RecipeAnvilRepair {
  enabled?: boolean;
  permission_suffix?: string;
  ingredient: VanillaItemsAdderCustomItem86;
  item: VanillaItemsAdderCustomItem89;
  [k: string]: unknown;
}
export interface RecipeSmithing {
  enabled?: boolean;
  permission_suffix?: string;
  base: VanillaItemsAdderCustomItem92;
  addition: VanillaItemsAdderCustomItem95;
  /**
   * The resulting item (can be a custom item or vanilla material) which will be merged into the first slot item.
   */
  result: {
    item: VanillaItemsAdderCustomItem98;
    amount?: number;
    /**
     * The attributes to be moved from the first item to the result item.
     */
    move_attributes?: {
      enchants?: boolean;
      durability?: boolean;
      custom_model_data?: boolean;
      material?: boolean;
      other_nbt?: boolean;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface LootBlock {
  enabled?: boolean;
  type?: VanillaItemsAdderCustomBlocks11;
  nbt?: Nbt;
  drop_only_first?: boolean;
  /**
   * List of the possible result drops and their chances
   */
  items?: {
    [k: string]: LootsResult;
  };
  biomes?: VanillaBiomes[];
  worlds?: (string | ("world" | "world_nether" | "world_the_end" | "!world_*" | "world_*"))[];
  /**
   * List of the possible result EXP drops and their chances
   */
  exp?: {
    [k: string]: ExpResult;
  };
  [k: string]: unknown;
}
// ...
export interface LootMobs {
  enabled?: boolean;
  type?: VanillaEntityTypes3;
  drop_only_first?: boolean;
  /**
   * List of the possible result drops and their chances
   */
  items?: {
    [k: string]: LootsResult1;
  };
  biomes?: VanillaBiomes[];
  /**
   * List of the possible result EXP drops and their chances
   */
  exp?: {
    [k: string]: ExpResult1;
  };
  nbt?: Nbt;
  /**
   * Special property to specify which metadata (Bukkit) data must match
   */
  metadata?: {
    [k: string]: MetadataMatcher;
  };
  [k: string]: unknown;
}
// ...
export interface LootFishing {
  enabled?: boolean;
  drop_only_first?: boolean;
  /**
   * List of the possible result drops and their chances
   */
  items?: {
    [k: string]: LootsResult2;
  };
  biomes?: VanillaBiomes[];
  /**
   * List of the possible result EXP drops and their chances
   */
  exp?: {
    [k: string]: ExpResult2;
  };
  [k: string]: unknown;
}
export interface LootsResult2 {
  item: VanillaItemsAdderCustomItem103;
  min_amount: number;
  max_amount: number;
  chance: number;
  ignore_fortune?: boolean;
  [k: string]: unknown;
}
export interface ExpResult2 {
  min_amount: number;
  max_amount: number;
  chance: number;
  [k: string]: unknown;
}
export interface WorldsPopulator {
  block: string;
  worlds: (string | ("world" | "world_nether" | "world_the_end"))[];
  replaceable_blocks?: VanillaBlocks3[];
  biomes?: VanillaBiomes[];
  chunk_chance: number;
  max_height: number;
  min_height: number;
  vein_blocks: number;
  chunk_veins: number;
  [k: string]: unknown;
}
export interface TreesPopulators {
  log: string;
  leaves: string;
  tree_type?:
    | "TREE"
    | "BIG_TREE"
    | "REDWOOD"
    | "TALL_REDWOOD"
    | "BIRCH"
    | "JUNGLE"
    | "SMALL_JUNGLE"
    | "COCOA_TREE"
    | "JUNGLE_BUSH"
    | "RED_MUSHROOM"
    | "BROWN_MUSHROOM"
    | "SWAMP"
    | "ACACIA"
    | "DARK_OAK"
    | "MEGA_REDWOOD"
    | "TALL_BIRCH"
    | "CHORUS_PLANT"
    | "CRIMSON_FUNGUS"
    | "WARPED_FUNGUS";
  worlds: (string | ("world" | "world_nether" | "world_the_end"))[];
  bottom_blocks?: VanillaBlocks4[];
  biomes?: VanillaBiomes[];
  chance: number;
  max_height: number;
  min_height: number;
  amount: number;
  iterations: number;
  [k: string]: unknown;
}
export interface SurfaceDecorators {
  block: string;
  worlds: (string | ("world" | "world_nether" | "world_the_end"))[];
  bottom_blocks?: VanillaBlocks5[];
  biomes?: VanillaBiomes[];
  chance: number;
  max_height: number;
  min_height: number;
  amount: number;
  [k: string]: unknown;
}
// ...
export interface CustomLiquid {
  color?: Colors;
  [k: string]: unknown;
}
// ...
export interface ArmorsRendering {
  [k: string]: ArmorsRenderingEntry;
}